requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.12.2
boto3==1.28.0
botocore==1.31.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
filelock==3.12.2
idna==3.4
jmespath==1.0.1
numpy==1.25.0
oscrypto==1.3.0
packaging==23.1
pandas==2.0.3
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
snowflake-connector-python==3.0.4
sortedcontainers==2.4.0
soupsieve==2.4.1
tabulate==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
